Correcting property types and added metaData info

+++ LeagueReplayFile/Models/Sections/GameDataSection.cs
@@ -3,6 +3,6 @@ namespace LeagueReplayFile.Models.Sections;
 public class GameDataSection : Section
 {
     public int ID { get; internal set; }
-    public int GameId { get; set; }
+    public long GameId { get; set; }
     public GameDataChunk Chunk { get; internal set; } = new GameDataChunk();
 }
